tests: Run the examples through lilypond
This gives us a basic syntax check for the lilypond formatter.

The tests verify that lilypond returns success and that lilypond's output
is error and warning free (technically checking for error is redundant
since 'error' should cause lilypond to return failure).

Note this causes a *big* drop in test performance.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>